+++ programs/ekr_client.c
@@ -111,7 +111,7 @@ conn_output(void *addr, void *buf, size_t length, uint8_t tos, uint8_t set_df)
 		usrsctp_freedumpbuffer(dump_buf);
 	}
 #ifdef _WIN32
-	if (send(*fdp, buf, length, 0) == SOCKET_ERROR) {
+	if (send(*fdp, buf, (int)length, 0) == SOCKET_ERROR) {
 		return (WSAGetLastError());
 #else
 	if (send(*fdp, buf, length, 0) < 0) {
